Make sure that udev device resources are released
When invoking udev_device_new_from_syspath(), a new udev device
with initial refcount 1 is created. This refcount must be decremented
to 0 to release the resources of the udev device.

Currently, udev_dev is first assigned to the return value of
udev_device_new_from_syspath(). Later, it is re-assigned to the return
value of udev_device_get_parent_with_subsystem_devtype(). As a result,
when we invoke udev_device_unref() to decrement the refcount, it refers
to the parent device instead of the device that was created.

According to the libudev docs, the device that is returned by
udev_device_get_parent_with_subsystem_devtype() is cleaned up when the
child device is cleaned up. In other words, invoking
udev_device_unref() on the child device should be sufficient here.

See libudev docs:
ftp://www.kernel.org/pub/linux/utils/kernel/hotplug/libudev/api-index-full.html