Increase retransmission timeout value in connectivity firmware
After observing frequent failures in pc-ble-driver, it became evident that the cause was often that the firmware did not get reponse from the PC side transport quickly enough, and started sending retransmissions. After some time this would often lead to failure. By increasing the timeout value in the connectivity, the issue seems to disappear, or at least appear so seldom that it has not been reproduced.