Fix problem related to control packet state transitions
The statemachine required too many criterias for changing from uninitialized to initialized state. According to the Bluetooth specfication (V4.2 [Vol 4, Part D], 8.X) it is sufficient to to receive a CONFIG_RESPONSE to transition to initialized state.
In addition there were missing mutex unlocks at reception of some control packets when in uninitized state. This would leave the state machine silently waiting for certain packets without responding to the incoming packets, or retransmitting according to what packet was expected.
Also added response to CONFIG packet in active state, as described in specification.