Use advisory locking of serial port on Linux/macOS
We want to prevent opening the same port multiple times. On Windows the
serial port is locked by the OS when opened. On Linux/macOS no locking
is done, so then we use the Boost.Interprocess advisory locking feature.

When trying to open a port that has been locked, it will now fail with
error "UART port /dev/ttyACMX is busy".