Change log functionality
DEBUG macro define was used as a macro that again invoked a small log api.

The DEBUG macro crash with both compiler defines and external
users log functionality.

This commit replaces that log functionality with a header only based
log macro (NRF_LOG).

This change in log functionality changed the timing in the application.
This revealed a bug where the state machine was incorrect when asserting
it in the tests.

To get better insight into the state machine state, a new state
STATE_NO_RESPONSE is introduced. Also the return value from
sd_rpc_open now returns more precise error codes.

A more aggressive locking of the state machine state is applied.
It may have an impact on for example BLE DFU operations.
This needs to be tested before this is merged into master.