Fix compilation on Linux, address bug discovered on Linux
This fixes gcc compilation errors that was discovered when compiling
on Linux. C++14 chrono::chrono_literals was used and there was a missing
include. Now hte code adheres to the C++11 standard.

Running the tests on Linux revealed a bug where a close operation
came in between state reset operations. This bug is fixed in this commit.

gcc compilation warnings are also fixed.

toString methods on ExitCriterias based classes are added to simplify
debugging.