Fix GH issue #112 Writing more than 20 bytes to peripheral crashes driver
When writing more than 20 bytes to a connected peripheral the connectivity firmware crash.
The issue is in the serialization codecs and is fixed in nRF5 SDK version 15 (SDKv15).

The fix requires the codecs for the connectivity firmware (connfw) to be updated.
To make this maintainable we align connfw for SoftDevice API v3 and v5 on SDKv15.
The build scripts and patch files are changed to make this happen.

The codecs on PC side is updated with codecs from SDKv15 since the codecs need to match the ones in connfw.

A test is implemented to verify that the issue is resolved.

Issues related to initialization of the H5Transport class and handling of unknown packets and state was discovered during fixing this issue.
This commit address these issues and reported linting issues in files affected.