Consolidate SoftDevice API v2, v3 and v5
We've had feedback from users telling us that it is not easy to
know what version of pc-ble-driver to use towards a given SoftDevice
API version.

We have also experienced that it is time consuming when maintaining
the pc-ble-driver project to have separate releases of pc-ble-driver for
different major versions.

This commit starts a strategy where we release all
SoftDevice API versions in every release of pc-ble-driver.

It also reuse parts of the codecs from the nRF5 SDK for SoftDevice
API v3 and onwards since parts of those are backwards compatible.