Automatically setup and run integration tests
This adds the ability to run the integration tests
automatically with the use of CMake targets.

The CMake targets program the development kits with the correct
connectivity firmware and then runs the relevant tests
for the development kit type and SoftDevice API version.
The tests are using the junit reporter as output.

It is still a limitation that std::cout and std::cerr is not
output in the junit reporter reports (or any other reporter).