Fix patches, change directory structure
This commit fixes patches that are applied to SDKv11 and SDKv15.
The contained empty lines without spaces that prevented the git command to apply parts of the patches.

The patches are also altered to support changing of connectivity firmware version number.

The changes to the directory structure make the files more intuitive to understand.