Port connectivity firmware build script to cmake (#136)
New build system for connectivity firmware

The previous way of building the connectivity firmware was a lot of manual steps supported by Windows .bat files and Cygwin/msys bash scripts.

This new build system uses directory structure conventions to automatically determine what and how to build the connectivity firmware.

Summarized it does the following:
- Downloads nRF5 SDK
- Applies patches to nRF5 SDK
- Searches for connectivity firmware projects
- Compiles connectivity firmware for combinations of SoftDevice API version, nRF family and transports (usb, 115k2, 1m)
- Merges connectivity with SoftDevice (by invoking mergehex)
- Creates DFU packages where relevant (by invoking nrfutil)

To override the SoftDevice version the developer can put a newer version in the relevant hex/sd_api_vX directory. The build system will pick that up and use that instead of the SoftDevice provided by the SDK.

New builds of the connectivity firmware is provided in this commit. They have not been tested.