Add check if adapter is already opened (#140)
* Add check if adapter is already opened

Previously checks if adapters was already opened or closed was not done. That lead to crashes. This commit adds logic to return status code NRF_ERROR_INVALID_STATE if the state is not valid when calling sd_rpc_open/sd_rpc_close.

* Update source code formatting rules

After internal review we have changed the formatting rules. Affected files are formatted according to clang-format.