Fix lock ordering
The lock for the state machine was done after currentState was retrieved. This could make the thread work on the wrong state exit criteria.
Thanks to @bencefr for spotting this.

Renamed variables to adhere to coding standard (removal of _ prefix) .

Changed the ordering of constructor member initializer the be aligned with declaration.