Add sequential compilation of connectivity
Previously compilation of targets was done in parallell
if the build system was given arguments to do that.
Since the source code is changed between different targets
(.in configuration files) that could lead to errors.

Also, certain compilers try to invoke more resources than
are available make the machine compiling connectivity
come close to a halt.