Fix thread issue in tests
The AdapterWrapper had a thread issue when destructuring an AdapterWrapper through a static std::vector with adapters. The static std::vector with adapters was again used by callbacks within the underlying Adapter.

This commit fixes that by changing the callbacks to be a static std::map within AdapterWrapper. The map is updated in constructor/destructor of AdapterWrapper, giving full control of the lifecycle.