Improve handling of ble_data_t buffers
Previously, there was a risk that buffer will be unregistered and
used later on. That resulted in test error when unexpected buffer
was received (null instead of valid buffer). Such situation was
occuring if adv_set_configure() call occured at the same time as
adv_set_terminated event was generated and advertising set was
active (adv_set_configure() was changing buffers).

Signed-off-by: Krzysztof Chruscinski <krzysztof.chruscinski@nordicsemi.no>