Transports: Use specific transports
This is an attempt at addressing a segfault issue on Linux where
the H5Transport do a double delete of the UartBoost transport.

The commit adds specific transports for the transports so
that it easier to understand the transport stack. To do this
the test_h5_transport.cpp is deleted. test_h5_transport is not used
in any test benches.

The destructors are cleaned up as they may be the cause of the segfault.