Tests: Install spdlog.dll for Windows (#274)
The spdlog port from vcpkg 2021.05.12 will build a spdlog.dll if
VCPKG_LIBRARY_LINKAGE is set to 'dynamic'; this happens for both the x86
and x64 Windows triplets. The result is that our test executable now has
a dependency on spdlog.dll, which wasn't being installed along the rest
of the files.

This commit fixes CMakeLists.txt so that it now properly installs
spdlog.dll. Interestingly, both the x64-linux and x64-osx triplets set
VCPKG_LIBRARY_LINKAGE to 'static', and thus don't need to install a
shared spdlog library.