Tests: Install spdlog.dll for Windows (#275)
The spdlog port from vcpkg 2021.05.12 will build a spdlog.dll if
VCPKG_LIBRARY_LINKAGE is set to 'dynamic'; this happens for both the x86
and x64 Windows triplets. The result is that our test executable now has
a dependency on spdlog.dll, which wasn't being installed along the rest
of the files.

This commit fixes CMakeLists.txt so that it now properly installs
spdlog.dll. Interestingly, both the x64-linux and x64-osx triplets set
VCPKG_LIBRARY_LINKAGE to 'static', and thus don't need to install a
shared spdlog library.

IMPORTANT: This commit is actually the same as
0c60de39d3cb33a9bb10eb294ac0ba3caff0ffc3. We need these changes here
in order to be able to build pc-ble-driver 4.1.4 using newer versions of
spdlog, such as those provided by vcpkg 2021.05.12 and above. Using that
commit directly would require us to do a new release of pc-ble-driver,
thoroughly test all the changes made since 4.1.4 up to that commit, etc.